+++ src/core/component/engines/vue/CHANGELOG.md
@@ -13,7 +13,7 @@ Changelog
 
 #### :house: Internal
 
-* Logging Vue errors and warnings via `core/log` module
+* Logging Vue errors and warnings via the `core/log` module
 
 ## v3.0.0-rc.37 (2020-07-20)